requirements.txt
가상환경에 pip install -r requirements.txt 해서 설치

absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
bidict==0.22.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
cycler==0.11.0
Django==3.2.19
et-xmlfile==1.1.0
filelock==3.12.0
Flask==2.1.3
Flask-SocketIO==5.3.4
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
gast==0.4.0
google-auth==2.19.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.2
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
image==1.5.33
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mediapipe==0.9.0.1
natsort==8.3.1
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
pyttsx3==2.90
pytz==2023.3
pywin32==306
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.1
scipy==1.7.3
simple-websocket @ file:///home/conda/feedstock_root/build_artifacts/simple-websocket_1681409243334/work
six==1.16.0
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.29.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685045523825/work
urllib3==1.26.16
Werkzeug==2.1.0
wincertstore==0.2
wrapt==1.15.0
wsproto @ file:///home/conda/feedstock_root/build_artifacts/wsproto_1661356345548/work
zipp==3.15.0
